+++ api/index.js
@@ -17,7 +17,7 @@ const path = require("path"),
     logPrefix = chalk
         .keyword(config.logger.consoleFgColor)
         .bgKeyword(config.logger.consoleBgColor)(` ${config.logger.consolePrefix} `),
-    logger = require("@logger").init(path.resolve(config.logger.logsPath), logPrefix);
+    logger = require("@logger").init(path.resolve(config.logger.logsPath), logPrefix, config.logger.level);
 
 logger.info("Express Template API v%s started%s. Path: %s", config.version, process.env.RUNNING_FOREVER ? " as daemon" : "", process.cwd());
 logger.info("Environment: " + config.env);