+++ facade/js/loader.js
@@ -1,0 +1,8 @@
+﻿export default {
+    show () {
+        document.getElementById("loader").style.display = "block";
+    },    
+    hide () {
+        document.getElementById("loader").style.display = "none";
+    }
+};