+++ facade/js/options.js
@@ -1,5 +1,0 @@
-export default {
-    env: process.env.NODE_ENV,
-    logEnabled: true,
-    debugEnabled: true
-};