+++ facade/js/pages/homepage.js
@@ -1,4 +1,4 @@
-import { log } from "core";
+import log from "logger";
 import $ from "jquery";
 import slideshow from "./../components/slideshow";